setting all child behaviour logic
