+++ src/environment.py
@@ -2,47 +2,6 @@ from entities import *
 
 import random
 
-class Direction:
-
-    @staticmethod
-    def north():
-        return -1, 0
-    
-    @staticmethod
-    def north_east():
-        return -1, 1
-    
-    @staticmethod
-    def east():
-        return 0, 1
-    
-    @staticmethod
-    def south_east():
-        return 1, 1
-    
-    @staticmethod
-    def south():
-        return 1, 0
-    
-    @staticmethod
-    def south_west():
-        return 1, -1
-    
-    @staticmethod
-    def west():
-        return 0, -1
-    
-    @staticmethod
-    def north_west():
-        return -1, -1
-
-directions = [Direction.north(), Direction.north_east(), Direction.east(), Direction.south_east(), 
-              Direction.south(), Direction.south_west(), Direction.west(), Direction.north_west()]
-
-def inside(env, x, y):
-    rows, cols = len(env), len(env[0])
-    return 0 <= x < rows and 0 <= y < cols
-
 def build_path(start, end, pi):
     if not end in pi: return []
